refact: migrated workflow to llm interface class
